New configuration where X-boxes don't get tagged frames
This achieves the same as the original configuration
(i.e., Broadcasts between SNAPs and the DHCP server work,
without inducing broadcast storms) but changes the configuraion
so that the X-boxes hang off access (not trunk) ports. This
is a bit more sane.

Broadcast storms are prevented by forcing all traffic towards
the SNAP switches to be tagged (apart from one link) which
can then be dropped at the front-end switches